It's time for some upgrades
Several Python, Ansible, Molecule, and Linux distro versions later,
it is time to bring this project up-to-speed.

A major challenge has been the ever-changing availability of certain JDKs
through SDKMAN. One day a specific version is available, the next day-- poof!
To get around this, a patch was implemented to allow `sdk install java` to
essentially download/install a default version. This should help address
the constant pain of fixing the CI tests.

Alpine is harder to support in light of the more recent Java releases.
The Java-based tests have been split out and the test Docker image
installs OpenJDK 11 through the `apk` package repository. This has been
noted in the `README.md`.

Molecule v3 upgrade has been tested on all scenarios.
