+++ molecule/alpine/converge.yml
@@ -14,14 +14,12 @@
         - candidate: groovy
           version: 1.8.9
       sdkman_install_packages:
-        - candidate: java
-          version: 8.0.232.hs-adpt
         - candidate: gradle
-          version: '4.6'
+          version: 6.8.1
         - candidate: gradle
-          version: 4.10.3
+          version: 6.7.1
       sdkman_defaults:
-        gradle: 4.10.3
+        gradle: 6.8.1
       sdkman_flush_caches_before:
         - temp
       sdkman_flush_caches_after:
@@ -30,10 +28,3 @@
         - temp
       sdkman_validate_ssl: true
       sdkman_offline_mode: true
-      sdkman_update_alternatives:
-        - candidate: java
-          name: java
-          link: /usr/bin/java
-        - candidate: java
-          name: javac
-          link: /usr/bin/javac
