+++ molecule/fedora/molecule.yml
@@ -3,33 +3,34 @@ dependency:
   name: galaxy
 driver:
   name: docker
-lint:
-  name: yamllint
+lint: |
+  set -e
+  yamllint .
+  flake8 tests
+  ansible-lint -x 208,106 .
 platforms:
-  - name: fedora28
-    image: fedora:27
+  - name: fedora33
+    image: fedora:33
     user: ${MOLECULE_USER}
-  - name: fedora27
-    image: fedora:27
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
+  - name: fedora32
+    image: fedora:32
     user: ${MOLECULE_USER}
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
 provisioner:
   name: ansible
   playbooks:
-    create: ../default/create.yml
     prepare: ../default/prepare.yml
-    converge: ../../tests/test.yml
-    destroy: ../default/destroy.yml
+    converge: ../default/converge.yml
   config_options:
     defaults:
       callback_whitelist: profile_tasks,timer
-  lint:
-    name: ansible-lint
 scenario:
   name: fedora
 verifier:
   name: testinfra
-  directory: ../../tests/
+  directory: ../../tests/default
   options:
     verbose: true
-  lint:
-    name: flake8
