+++ molecule/ubuntu/molecule.yml
@@ -3,36 +3,44 @@ dependency:
   name: galaxy
 driver:
   name: docker
-lint:
-  name: yamllint
+lint: |
+  set -e
+  yamllint .
+  flake8 tests
+  ansible-lint -x 208,106 .
 platforms:
+  - name: focal
+    image: ubuntu:focal
+    user: ${MOLECULE_USER}
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
   - name: bionic
     image: ubuntu:bionic
     user: ${MOLECULE_USER}
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
   - name: xenial
     image: ubuntu:xenial
     user: ${MOLECULE_USER}
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
   - name: trusty
     image: ubuntu:trusty
     user: ${MOLECULE_USER}
+    dockerfile: ../default/Dockerfile.j2
+    tty: true
 provisioner:
   name: ansible
   playbooks:
-    create: ../default/create.yml
     prepare: ../default/prepare.yml
-    converge: ../../tests/test.yml
-    destroy: ../default/destroy.yml
+    converge: ../default/converge.yml
   config_options:
     defaults:
       callback_whitelist: profile_tasks,timer
-  lint:
-    name: ansible-lint
 scenario:
   name: ubuntu
 verifier:
   name: testinfra
-  directory: ../../tests/
+  directory: ../../tests/default
   options:
     verbose: true
-  lint:
-    name: flake8
