+++ tests/alpine/test_alpine.py
@@ -1,0 +1,1 @@
+from ..shared import *
