+++ tests/test-requirements.txt
@@ -1,3 +1,5 @@
-ansible>=2.8,<2.9
+ansible
+ansible-lint
+flake8
 molecule[docker]
-testinfra>3.0
+pytest-testinfra
